+++ Gruntfile.js
@@ -15,24 +15,18 @@ module.exports = function exports(grunt) {
         DEBUG: 'price-finder*',
       },
     },
-    jasmine_nodejs: {
-      options: {
-        specNameSuffix: 'test.js',
-        reporters: {
-          console: {
-            colors: true,
-            cleanStack: true,
-            verbosity: true,
-            listStyle: 'indent',
-            activity: false,
-          },
-        },
-      },
+    mochaTest: {
       unit: {
-        specs: ['test/unit/**'],
+        src: ['test/unit/**/*test.js'],
       },
       e2e: {
-        specs: ['test/e2e/**'],
+        options: {
+          reporter: 'list',
+
+          // set the timeout for each test to 60 seconds (crazy! but necessary it seems)
+          timeout: 60000,
+        },
+        src: ['test/e2e/**/*test.js'],
       },
     },
   });
@@ -40,6 +34,6 @@ module.exports = function exports(grunt) {
   // load all the grunt tasks at once
   require('load-grunt-tasks')(grunt);
 
-  grunt.registerTask('test', ['eslint', 'jasmine_nodejs:unit']);
-  grunt.registerTask('e2e', ['env:e2e', 'jasmine_nodejs:e2e']);
+  grunt.registerTask('test', ['eslint', 'mochaTest:unit']);
+  grunt.registerTask('e2e', ['env:e2e', 'mochaTest:e2e']);
 };