+++ test/e2e/greenman-gaming-uris-test.js
@@ -7,9 +7,9 @@ const priceFinder = testHelper.priceFinder;
 const verifyPrice = testHelper.verifyPrice;
 const verifyItemDetails = testHelper.verifyItemDetails;
 
-describe('price-finder for Snapdeal Store URIs', () => {
-  describe('testing item', () => {
-    const uri = 'http://www.snapdeal.com/product/iphone-6s-16gb/663413326062';
+describe('price-finder for GreenmanGaming Store URIs', () => {
+  describe('testing Homefront: The Revolution item', () => {
+    const uri = 'http://www.greenmangaming.com/s/in/en/pc/games/action/homefront-revolution/';
 
     it('should respond with a price for findItemPrice()', (done) => {
       priceFinder.findItemPrice(uri, (err, price) => {
@@ -22,7 +22,7 @@ describe('price-finder for Snapdeal Store URIs', () => {
     it('should respond with a price, and the right category and name for findItemDetails()', (done) => {
       priceFinder.findItemDetails(uri, (err, itemDetails) => {
         should(err).be.null();
-        verifyItemDetails(itemDetails, 'iPhone 6s 16GB', 'Mobile Phones');
+        verifyItemDetails(itemDetails, 'Homefront: The Revolution (NA)', 'Video Games');
         done();
       });
     });