+++ test/e2e/priceminister-uris-test.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const should = require('should');
 const testHelper = require('./test-helper');
 
 const priceFinder = testHelper.priceFinder;
@@ -9,12 +10,12 @@ const verifyItemDetails = testHelper.verifyItemDetails;
 describe('price-finder for PriceMinister Store URIs', () => {
   // Video Games
   describe('testing a Video Game item', () => {
-    // Grand Theft Auto IV sur PS3
-    const uri = 'http://www.priceminister.com/mfp/2001701/gta-4-jeu-video#pid=63080574';
+    // Call Of Duty : Black Ops III sur PS4
+    const uri = 'http://www.priceminister.com/offer/buy/999031035/call-of-duty-black-ops-iii.html';
 
     it('should respond with a price for findItemPrice()', (done) => {
       priceFinder.findItemPrice(uri, (err, price) => {
-        expect(err).toBeNull();
+        should(err).be.null();
         verifyPrice(price);
         done();
       });
@@ -22,8 +23,8 @@ describe('price-finder for PriceMinister Store URIs', () => {
 
     it('should respond with a price, and the right category and name for findItemDetails()', (done) => {
       priceFinder.findItemDetails(uri, (err, itemDetails) => {
-        expect(err).toBeNull();
-        verifyItemDetails(itemDetails, 'Gta 4 jeu video', 'Video Games');
+        should(err).be.null();
+        verifyItemDetails(itemDetails, 'Call of duty black ops iii', 'Video Games');
         done();
       });
     });